requirements.txt
Django==2.0.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
python-dateutil==2.6.1
virtualenv==15.1.0

